Added JsLiveEditor Rules Scripted Condition / Consequence functional (still need to stabilize) getDynamicObject() in Controllable and getValue in parameter, allows to get realtime value of parameters without having to recreate js environment

+++ src/LGML/Source/MIDIController.cpp
@@ -71,7 +71,7 @@ void MIDIController::onContainerParameterChanged(Parameter * p){
 void MIDIController::buildLocalEnv(){
     DynamicObject obj;
     obj.setMethod("sendCC", sendCC);
-    obj.setProperty(ptrIdentifier, this);
+    obj.setProperty(ptrIdentifier, (int64)this);
     obj.setMethod("sendNoteOnFor", sendNoteOnFor);
     setLocalNamespace(obj);
